Hacer búsqueda por Idfacultad en Facultades (evidencias)

+++ index.php
@@ -7,6 +7,9 @@ global $titleDocument;
 $titleDocument = 'Página de inicio';
 include('./controller/server.php');
 
+$schoolCode = isset($_GET['txtSchoolCode']) ? $_GET['txtSchoolCode'] : '';
+$schoolCode_error = true;
+
 ?>
 
 <!DOCTYPE html>
@@ -19,10 +22,29 @@ include('./controller/server.php');
         <div class="container-fluid">
             <div class="row">
                 <div class="col-md-12">
+                    <h2 class="pull-left">Detalles de Facultades</h2>
                     <div class="mt-5 mb-3 clearfix">
-                        <h2 class="pull-left">Detalles de Facultades</h2>
-                        
-                        <a href="./view/school/create.php" class="btn btn-success pull-right"><i class="fa fa-plus"></i> Agregar una Facultad</a>
+                        <div class="row">
+                            <div class="col-6">
+                                <div class="row">
+                                    <div class="col mt-3">
+                                        <form>
+                                            <div class="form-group">
+                                                <label>Consultar por ID Facultad en la tabla</label>
+                                                <input type="number" name="txtSchoolCode" class="form-control" autocomplete="off" value="<?php echo $schoolCode ?>">
+                                            </div>
+                                        </form>
+                                    </div>
+                                </div>
+                            </div>
+                            <div class="col-6">
+                                <div class="row">
+                                    <div class="col mt-5">
+                                        <a href="./view/school/create.php" class="btn btn-success pull-right"><i class="fa fa-plus"></i> Agregar una Facultad</a>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
                     </div>
                     <?php
                     // Include config file
@@ -38,7 +60,7 @@ include('./controller/server.php');
                         echo '<table class="table table-bordered table-striped">';
                         echo "<thead>";
                         echo "<tr>";
-                        echo "<th>Código</th>";
+                        echo "<th>ID Facultad</th>";
                         echo "<th>Nombre Facultad</th>";
                         echo "<th>Nombre del Decano</th>";
                         echo "<th>Nombre IES</th>";
@@ -47,20 +69,45 @@ include('./controller/server.php');
                         echo "</thead>";
                         echo "<tbody>";
                         foreach ($row as $res) {
-                            echo "<tr>";
-                            echo "<td>" . $res['Idfacultades'] . "</td>";
-                            echo "<td>" . $res['Nombre'] . "</td>";
-                            echo "<td>" . $res['Decano'] . "</td>";
-                            echo "<td>" . $res['Ies_nombre'] . "</td>";
-                            echo "<td>";
-                            echo '<a href="./view/school/read.php?id=' . $res['Idfacultades'] . '" class="mr-3" title="View Record" data-toggle="tooltip"><span class="fa fa-eye"></span></a>';
-                            echo '<a href="./view/school/update.php?id=' . $res['Idfacultades'] . '" class="mr-3" title="Update Record" data-toggle="tooltip"><span class="fa fa-pencil"></span></a>';
-                            echo '<a href="./view/school/delete.php?id=' . $res['Idfacultades'] . '" title="Delete Record" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
-                            echo "</td>";
-                            echo "</tr>";
+                            if ($schoolCode != '') {
+                                if ($schoolCode == $res['Idfacultades']) {
+                                    echo "<tr>";
+                                    echo "<td>" . $res['Idfacultades'] . "</td>";
+                                    echo "<td>" . $res['Nombre'] . "</td>";
+                                    echo "<td>" . $res['Decano'] . "</td>";
+                                    echo "<td>" . $res['Ies_nombre'] . "</td>";
+                                    echo "<td>";
+                                    echo '<a href="./view/school/read.php?id=' . $res['Idfacultades'] . '" class="mr-3" title="View Record" data-toggle="tooltip"><span class="fa fa-eye"></span></a>';
+                                    echo '<a href="./view/school/update.php?id=' . $res['Idfacultades'] . '" class="mr-3" title="Update Record" data-toggle="tooltip"><span class="fa fa-pencil"></span></a>';
+                                    echo '<a href="./view/school/delete.php?id=' . $res['Idfacultades'] . '" title="Delete Record" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
+                                    echo "</td>";
+                                    echo "</tr>";
+                                    if ($res['Idfacultades'] === null) {
+                                        $schoolCode_error = true;
+                                    } else {
+                                        $schoolCode_error = false;
+                                    }
+                                }
+                            } else {
+                                echo "<tr>";
+                                echo "<td>" . $res['Idfacultades'] . "</td>";
+                                echo "<td>" . $res['Nombre'] . "</td>";
+                                echo "<td>" . $res['Decano'] . "</td>";
+                                echo "<td>" . $res['Ies_nombre'] . "</td>";
+                                echo "<td>";
+                                echo '<a href="./view/school/read.php?id=' . $res['Idfacultades'] . '" class="mr-3" title="View Record" data-toggle="tooltip"><span class="fa fa-eye"></span></a>';
+                                echo '<a href="./view/school/update.php?id=' . $res['Idfacultades'] . '" class="mr-3" title="Update Record" data-toggle="tooltip"><span class="fa fa-pencil"></span></a>';
+                                echo '<a href="./view/school/delete.php?id=' . $res['Idfacultades'] . '" title="Delete Record" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
+                                echo "</td>";
+                                echo "</tr>";
+                                $schoolCode_error = false;
+                            }
                         }
                         echo "</tbody>";
                         echo "</table>";
+                        if ($schoolCode_error) {
+                            echo '<div class="alert alert-danger"><em>No se encontraron resultados con el ID Facultad </em><span class="font-weight-bold">', $schoolCode, '</span></div>';
+                        }
                     } else {
-                        echo '<div class="alert alert-danger"><em>No records were found.</em></div>';
+                        echo '<div class="alert alert-danger"><em>No existe información en la base de datos.</em></div>';
                     }
